project/requirements.txt
networkx>=1.11
PyYAML>=3.11
tqdm>=4.10.0
pygraphviz>=1.3.1
numpy>=1.11.2
scipy>=0.18.1